+++ CHANGELOG.md
@@ -40,6 +40,9 @@ The format is based on [Keep a Changelog].
   (`mouse-1`) selects the candidate, and right click (`mouse-3`)
   inserts the candidate, just like `RET` and `TAB`, respectively. See
   [#113] and [#118].
+* The user option `selectrum-extend-current-candidate-highlight`
+  determines whether to extend the highlighting of the current
+  candidate until the margin (the default is nil). See [#208].
 
 ### Enhancements
 * `selectrum-insert-current-candidate` no longer adds inserted
@@ -189,6 +192,7 @@ The format is based on [Keep a Changelog].
 [#193]: https://github.com/raxod502/selectrum/pull/193
 [#197]: https://github.com/raxod502/selectrum/pull/197
 [#198]: https://github.com/raxod502/selectrum/pull/198
+[#208]: https://github.com/raxod502/selectrum/pull/208
 [#212]: https://github.com/raxod502/selectrum/issues/212
 [#213]: https://github.com/raxod502/selectrum/pull/213